sdlc_files/requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
PyPDF2==3.0.1
llama-cpp-python==0.3.9

# Frontend dependencies
streamlit==1.28.1
requests==2.31.0

# Additional utilities
python-dotenv==1.0.0
